Move colored output test to span module
